+++ tests/Browser/SitesTest.php
@@ -31,28 +31,6 @@ class SitesTest extends DuskTestCase
      * @test
      * @group sites
      */
-    public function it_can_navigate_in_pagination()
-    {
-        $client = factory(Client::class, 1)->create()->first();
-        $sites = factory(Site::class, 50)->create(['client_id' => $client->id]);
-        $site = Site::find(26);
-
-        $this->browse(function (Browser $browser) use ($sites, $site) {
-            $browser->loginAs(1)
-                ->visit(new Sites())
-                ->waitFor('@table')
-                ->assertSee('Next')
-                ->assertDontSeeIn('@table', $site->name)
-                ->clickLink('Next')
-                ->pause(1000)
-                ->assertSeeIn('@table', $site->name);
-        });
-    }
-
-    /**
-     * @test
-     * @group sites
-     */
     public function it_can_create_a_site()
     {
         $client = factory(Client::class, 1)->create()->first();
@@ -70,6 +48,7 @@ class SitesTest extends DuskTestCase
                 ->waitForText('The client id field is required.')
                 ->select('@inputClient', (string)$client->id)
                 ->press('Save')
+                ->waitUntilMissing('@modalForm')
                 ->waitFor('@table')
                 ->assertSeeIn('@table', 'My blog');
         });
@@ -151,6 +130,31 @@ class SitesTest extends DuskTestCase
                 ->waitFor('@modalHistory')
                 ->assertSeeIn('@modalHistory', $site->name)
                 ->assertSeeIn('@modalHistory', $original_name);
+        });
+    }
+
+    /**
+     * @test
+     * @group sites
+     */
+    public function it_can_navigate_in_pagination()
+    {
+        $this->runDatabaseMigrations();
+
+        $client = factory(Client::class, 1)->create()->first();
+        $sites = factory(Site::class, 50)->create(['client_id' => $client->id]);
+
+        $this->browse(function (Browser $browser) {
+            $browser->loginAs(1)
+                ->visit(new Sites())
+                ->waitFor('@table')
+                ->assertSee('Next');
+
+            $value = $browser->text('table.table tbody tr td');
+
+            $browser->clickLink('Next')
+                    ->pause(1000)
+                    ->assertDontSeeIn('@table', $value);
         });
     }
 }